refactor: 홈 API·UI를 features/home으로 이동 (FSD 구조 개선) (#43)
- entities/home/ 삭제 후 features/home/으로 이동
  - api/homeApi.ts, api/queries.ts, model/home.types.ts
- widgets/home/ 하위 UI 컴포넌트를 features/home/ui/로 이동
  - HomeBanner, DiaryWriteButton, RandomSentenceBanner, useInfiniteSlider
- widgets/home/index.ts는 HomeDiarySection만 export하도록 정리
- HomeDiarySection은 features/home + entities/diary를 조합하는 widget으로 유지
