--- src/entities/home/index.ts
@@ -1,3 +1,0 @@
-export { fetchHomeRandom, fetchHomeSummary } from "./api/homeApi";
-export { homeKeys, useHomeRandomQuery, useHomeSummaryQuery } from "./api/queries";
-export type { HomeSummaryResponse, MonthlyDiaryResponse, QuoteResponse } from "./model/home.types";
