+++ src/features/home/model/useInfiniteSlider.ts
