+++ src/features/home/ui/HomeBanner.tsx
